Extract feature collection map creation into a helper
