+++ moped-editor/src/views/projects/projectView/ProjectComponents/utils/map.js
